requirements.txt
fastapi
uvicorn
rembg
torch
torchvision
pillow
pillow-avif-plugin
python-dotenv
python-multipart
requests